Worker cleanup

+++ kolejka/worker/stage0.py
@@ -2,6 +2,7 @@
 
 import copy
 import glob
+import logging
 import math
 import os
 import shutil
@@ -10,16 +11,23 @@ import sys
 import tempfile
 import uuid
 
-from kolejka.common.settings import OBSERVER_SOCKET, TASK_SPEC, WORKER_LIMITS
-from kolejka.common import silent_call
+from kolejka.common.settings import OBSERVER_SOCKET, TASK_SPEC, RESULT_SPEC, WORKER_HOSTNAME, WORKER_REPOSITORY
 from kolejka.common import KolejkaTask, KolejkaResult
+
+def silent_call(*args, **kwargs):
+    kwargs['stdin'] = kwargs.get('stdin', subprocess.DEVNULL)
+    kwargs['stdout'] = kwargs.get('stderr', subprocess.DEVNULL)
+    kwargs['stderr'] = kwargs.get('stdout', subprocess.DEVNULL)
+    return subprocess.call(*args, **kwargs)
 
 def stage0(task_path, result_path, temp_path=None):
     task = KolejkaTask(task_path)
-#    if task.id is None:
-#        task.id = uuid.uuid4().hex
-#        task.commit()
-    print(task.id, task.args, task.files)
+    if task.id is None:
+        task.id = uuid.uuid4().hex
+    assert task.image is not None
+    assert len(task.args) > 0
+    assert task.files.is_local
+
     docker_task = 'kolejka_task_{}'.format(task.id)
 
     docker_cleanup  = [
@@ -31,13 +39,20 @@ def stage0(task_path, result_path, temp_path=None):
         os.makedirs(os.path.join(jailed_path, 'task'), exist_ok=True)
         os.makedirs(os.path.join(jailed_path, 'result'), exist_ok=True)
         os.makedirs(os.path.join(jailed_path, 'dist'), exist_ok=True)
+        jailed = KolejkaTask(os.path.join(jailed_path, 'task'))
+        jailed.load(task.dump())
+        jailed.files.clear()
         volumes = list()
+        assert os.path.exists(OBSERVER_SOCKET)
         volumes.append((OBSERVER_SOCKET, OBSERVER_SOCKET, 'rw'))
         volumes.append((os.path.join(jailed_path, 'result'), '/opt/kolejka/result', 'rw'))
         for key, val in task.files.items():
             if key != TASK_SPEC:
                 volumes.append((os.path.join(task.path, val.path), os.path.join('/opt/kolejka/task', key), 'ro'))
-        volumes.append((task.spec_path, os.path.join('/opt/kolejka/task', TASK_SPEC), 'ro'))
+                jailed.files.add(key)
+        jailed.files.add(TASK_SPEC)
+        jailed.commit()
+        volumes.append((jailed.spec_path, os.path.join('/opt/kolejka/task', TASK_SPEC), 'ro'))
         for spath in [ os.path.dirname(__file__) ]:
             stage1 = os.path.join(spath, 'stage1.sh')
             if os.path.isfile(stage1):
@@ -49,17 +64,12 @@ def stage0(task_path, result_path, temp_path=None):
                 if len(packs) > 0:
                     pack = sorted(packs)[-1]
                     volumes.append((pack, '/opt/kolejka/dist/'+package+'-1-py3-none-any.whl', 'ro'))
-        print(volumes)
 
         docker_call  = [ 'docker', 'run' ]
-#        docker_call += [ '--cgroup-parent', docker_task ]
-#        docker_call += [ '--cpus', str(task.cpus) ]
-#        for device in task.devices :
-#            docker_call += [ '--device', device ]
         docker_call += [ '--entrypoint', '/opt/kolejka/stage1.sh' ]
         for key, val in task.environment.items():
             docker_call += [ '--env', '{}={}'.format(key, val) ]
-        docker_call += [ '--hostname', 'kolejka' ]
+        docker_call += [ '--hostname', WORKER_HOSTNAME ]
         docker_call += [ '--init' ]
         if task.limits.memory is not None:
             docker_call += [ '--memory', str(task.limits.memory) ]
@@ -70,37 +80,37 @@ def stage0(task_path, result_path, temp_path=None):
             docker_call += [ '--pids-limit', str(task.limits.pids) ]
         if task.limits.time is not None:
             docker_call += [ '--stop-timeout', str(int(math.ceil(task.limits.time.total_seconds()))) ]
-        assert os.path.exists(OBSERVER_SOCKET)
         for v in volumes:
             docker_call += [ '--volume', '{}:{}:{}'.format(os.path.realpath(v[0]), v[1], v[2]) ]
         docker_call += [ '--workdir', '/opt/kolejka' ]
-        docker_call += [ 'kolejka.matinf.uj.edu.pl/' + task.image ]
+        docker_call += [ WORKER_REPOSITORY.rstrip('/')+ '/' + task.image ]
         docker_call += [ '--debug' ]
 
         for docker_clean in docker_cleanup:
             silent_call(docker_clean)
 
-        print(docker_call)
-
         subprocess.check_call(docker_call)
 
-
-        result = KolejkaResult(os.path.join(jailed_path, 'result'))
+        jailed = KolejkaResult(os.path.join(jailed_path, 'result'))
 
         for docker_clean in docker_cleanup:
             silent_call(docker_clean)
 
-        shutil.rmtree(result_path)
+        if os.path.exists(result_path):
+            shutil.rmtree(result_path)
         os.makedirs(result_path, exist_ok=True)
-        reresult = KolejkaResult(result_path)
-        dum = result.dump()
-        del dum['files']
-        reresult.load(dum)
-        for key, val in result.files.items():
-            print(key)
-            with open(os.path.join(result.path, val.path), 'r') as file_file:
-                print(file_file.read())
-            os.makedirs(os.path.dirname(os.path.join(reresult.path, key)), exist_ok=True)
-            shutil.move(os.path.join(result.path, val.path), os.path.join(reresult.path, key))
-            reresult.files.add(key)
-        reresult.commit()
+        result = KolejkaResult(result_path)
+        result.load(jailed.dump())
+        result.files.clear()
+        for key, val in jailed.files.items():
+            if key != RESULT_SPEC:
+                print(key)
+                with open(os.path.join(jailed.path, val.path), 'r') as file_file:
+                    print(file_file.read())
+                os.makedirs(os.path.dirname(os.path.join(result.path, key)), exist_ok=True)
+                shutil.move(os.path.join(jailed.path, val.path), os.path.join(result.path, key))
+                result.files.add(key)
+        result.commit()
+        print(RESULT_SPEC)
+        with open(result.spec_path, 'r') as file_file:
+            print(file_file.read())
